move the lint.rb file to lib directory
